+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN6MC8M3TE3WDXVQXCT78346"
+        "01KN6QXCKKCA49RD3X2P8GB40T"
     )
     r"""The unique identifier of the data source"""