+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KN6MC86BDBQFD3KCG5F6898Z"
+    id: Optional[str] = "01KN6QXC5H8P6EFEVT7BGW31TW"
 
     description: Optional[str] = None