+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN6MC8M8DPXQ9F4Y6CNXRNR1"
+        "01KN6QXCKRZ317JWJQT5T2T2GP"
     )
     r"""The unique identifier of the data source"""